+++ src/index.ts
@@ -2,10 +2,10 @@
 export * from './foundation'
 
 /* Components */
+export * from './components/Icon'
+export * from './components/Button'
 export * from './components/Banner'
 export * from './components/Divider'
-export * from './components/Icon'
-export * from './components/Forms/Switch'
 export * from './components/Text'
 export * from './components/Tabs'
 export * from './components/TabItem'
@@ -24,8 +24,8 @@ export * from './components/Avatars/CheckableAvatar'
 export * from './components/TagBadge'
 export * from './components/Status'
 export * from './components/KeyValueListItem'
-export * from './components/Forms/Button'
 export * from './components/Forms/Radio'
+export * from './components/Forms/Switch'
 export * from './components/Forms/Checkbox'
 export * from './components/Forms/SegmentedControl'
 export * from './components/Forms/Inputs/Select'
